backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.2.1
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.1
